Move _get_some_code method to InteractiveShell object; improve docstring.
